Write initial README content
Create the new repository's README.md boilerplate content which
explains what the project is and what the (expected) content of the
repository is going to be.

As this is the first commit related to content, this is expected to be
simply a basic boilerplate which will be built upon later.