Create Terraform DigitalOcean droplet resource
Creates the initial boilerplate required to interact with the
DigitalOcean API and then builds upon this by implementing the
DigitalOcean droplet resource itself.

This is done in a single commit as this is the first content-related
commit for Terraform, so reverting just the droplet resource should
not be of any value at this point.

This may come back to bite me of course though.